requirements.txt
blinker==1.9.0
click==8.1.8
dateutils==0.6.12
Flask==3.1.0
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
SQLAlchemy==2.0.39
typing_extensions==4.12.2
Werkzeug==3.1.3
